move out lot of form functionality into agent form, move some schemas to character.ts
